requirements.txt
datacube==1.8.3
numpy
pandas
pygeoapi
pyproj
PyYAML
rasterio
scipy
